# Removed 'post_status' property from redirects data