@dkotter 2023-06-30
> I suggest keeping post_status in the result for queries with post_status=> any or only for CLI commands.

Okay, that was the biggest thing to make sure of here, that there was no downside to removing this check. Sounds like we may need to keep this at least in `WP-CLI` contexts or maybe just anytime the query has `post_status => 'any'`

> This pull request does not resolve the issue entirely but is an excellent first foot for performance improvement.

Yes, there's other things we could look at doing but seemed like this was by far the easiest thing to work on initially. We may come back to some of the other things that were discussed on #24