@sksaju 2023-07-05
> @sksaju As you can see, I discussed with @dkotter ([#326 (comment)](https://github.com/10up/safe-redirect-manager/pull/326#issuecomment-1615160076)); we decided to include `post_status` in the result if `post_status` can have more that one value which is possible if querying with `post_status => any` or `post_status => { array of post status }`.
> 
> Can you make this change to the pull request? Let me know if you have any questions.

Yeah, I've updated the PR. Thanks