@sksaju 2023-07-19
> @sksaju, do you have any questions about my recent [comment](https://github.com/10up/safe-redirect-manager/pull/326#pullrequestreview-1514372652)?

Hi @ravinderk, Thank you for your feedback. I've implemented the check you mentioned and tested it locally. Please review the latest commit and let me know if there's anything else I can do. Thanks.